requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.42.0,<0.43.0

python-dotenv>=1.0.0,<2.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg[binary]>=3.1.0,<4.0.0

requests>=2.32.0,<3.0.0
dnspython>=2.6.0,<3.0.0
python-nmap>=0.7.1,<0.8.0
python-whois>=0.9.5,<1.0.0

# Provides: `from Wappalyzer import Wappalyzer, WebPage`
# Note: this package is older/unmaintained, but matches the repo's import.
